Add thumbnail updating functionality; USE AT OWN RISK
It can and will overwrite old jpegs you output if you open a file.
Be careful.

I had to change many types to get things to line up right, and I added a
new thread for processing the thumbnails without blocking the UI.

+++ filmulator-gui/database/sqlInsertion.cpp
@@ -127,7 +127,7 @@ void createNewProfile(const QString fileHash,
     query.bindValue(24, defaultQuery.value(24).toFloat());//highlightsY
     query.bindValue(25, defaultQuery.value(25).toInt()  );//highlightRecovery
     query.bindValue(26, defaultQuery.value(26).toInt()  );//caEnabled
-    double temp, tint;
+    float temp, tint;
     optimizeWBMults(absoluteFilePath.toStdString(), temp, tint);
     query.bindValue(27, temp);//defaultQuery.value(27).toFloat());//temperature
     query.bindValue(28, tint);//defaultQuery.value(28).toFloat());//tint
